Fixed npcs, removed a game constant and added keylistener class
* Fixed npcs layout.
* Removed arraylist of npc types in game constant.
* Added keylistener class.
